edge-health: support extendable check plugin
Implement the plugin interface of their own strategy as a plugin factory to tell superedge to add the plugins.
Register the plugins as a parameter by calling NewEdgeHealthCommand.
The inspiration is from kube-scheduler out-of-tree plugin registry. The project scheduler-plugin is based on this cool function. It adds new scheduler plugins without modifying any source codes in kube-scheduler.

Fixes #316

Signed-off-by: JaneLiuL <janeliul1220@gmail.com>